feat(BuySell): splits up coinify details modal into trade and kyc components
