+++ packages/blockchain-wallet-v4-frontend/src/modals/CoinifyTradeDetails/index.js
@@ -10,66 +10,24 @@ import { OrderDetailsTable, OrderDetailsRow } from 'components/BuySell/OrderDeta
 import { tradeDetails, statusHelper, bodyStatusHelper } from 'services/CoinifyService'
 import { spacing } from 'services/StyleService'
 
-const ButtonRow = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-end;
-  margin-top: 20px;
-`
-
-const renderDetailsRow = (id, message, value, color) => (
-  <OrderDetailsRow>
-    <Text size='13px' weight={300}><FormattedMessage id={id} defaultMessage={message} /></Text>
-    <Text size='13px' weight={300} color={color}>{value}</Text>
-  </OrderDetailsRow>
-)
+import Trade from './Trade'
+import Kyc from './KYC'
 
 class CoinifyTradeDetails extends React.PureComponent {
   render () {
     console.log('CoinifyTradeDetails', this.props)
-    const headerStatus = statusHelper(this.props.trade.state)
-    const bodyStatus = bodyStatusHelper(this.props.trade.state, this.props.trade.isBuy)
+
     const { trade, status } = this.props
-    const details = tradeDetails.renderDetails(trade)
+
+    const renderComponent = (trade) => (
+      trade.constructor.name === 'Trade'
+        ? <Trade trade={trade} close={this.props.close} />
+        : <Kyc status={status} />
+    )
 
     return (
       <Modal size='large' position={this.props.position} total={this.props.total}>
-        <ModalHeader onClose={this.props.close}>
-          <Text color={headerStatus.color}>
-            { trade.isBuy ? `Buy Order` : 'Sell Order' } {headerStatus.text}
-          </Text>
-        </ModalHeader>
-        <ModalBody>
-          <Text size='13px' weight={300}>
-            { bodyStatus.text }
-          </Text>
-          <Text style={spacing('pt-5')} size='13px' weight={300}>
-            <FormattedMessage id='order_details.trade_id' defaultMessage={`Your order ID is: CNY-{id}`} values={{ id: trade.id }} />
-          </Text>
-          <OrderDetailsTable style={spacing('mt-10')}>
-            {renderDetailsRow(
-              'order_details.amount_to_purchase',
-              trade.isBuy ? 'BTC Purchased' : 'BTC Sold',
-              details.firstRow
-            )}
-            {renderDetailsRow(
-              'order_details.trading_fee',
-              'Trading Fee',
-              details.fee
-            )}
-            {renderDetailsRow(
-              'order_details.total_cost',
-              trade.isBuy ? 'Total Cost' : 'Total To Be Received',
-              details.total,
-              'success'
-            )}
-          </OrderDetailsTable>
-          <ButtonRow>
-            <Button width='100px' onClick={this.props.close} nature='primary'>
-              <FormattedMessage id='close' defaultMessage='Close' />
-            </Button>
-          </ButtonRow>
-        </ModalBody>
+        { renderComponent(trade) }
       </Modal>
     )
   }
